Move function for getting url params object to utils
